Refactor getDataFromRelay in WalletViewController.
